big ass overhaul
There may be an inconsistency here as the default dropdown Apple uses
at the top of the page is a <label> tag, and so is my background
switcher… so there is a bug if you use both the regular dropdown and
then try to evoke background switcher - it will just open regular
dropdown again. Anyway it’s late and I’ve been entertaining myself for
an entire day with this. I think I’ve had my fun out of it. Night.