Chrome Sync Set
What a day! I’ve managed to get the value to save in
chrome.storage.sync.set and i can successfully retrieve it. The next
step will be to replace the bg var with the output of
chrome.storage.sync.get “background” key but this was proving much
harder than I first thought to do.