Saves the background
Utilising the power from chrome.storage.sync

Ability to save text strings without a database has enabled me to
create a save feature so that when you quit Chrome and come back to it
your background choice is retained. The variable “bg” now has the value
of the data stored in the “background” key in chrome.storage.sync. Have
not tested but in theory if you are signed into your Google Account
this setting will be retained on different computer too.